requirements.txt
pandas===2.2.2
numpy===1.26.4
opencv-python===4.10.0.84
matplotlib===3.9.2
tqdm===4.66.5
lxml===5.2.1
pytorch===2.5.1
pytorch-cuda===12.4
torchinfo===1.8.0
torchvision===0.20.1